package: qemu: Make dependency to nettle explicit
For the fli4l router distribution we had problems building 'qemu', some
builds failed complaining when checking dynamic linking of the package:

>>> qemu 2.12.1 Error: Library libnettle.so.6 referenced by usr/libexec/qemu-bridge-helper belongs to package nettle which is not a dependency of qemu

https://web.nettworks.org/ci/job/fli4l/job/buildroot/job/4.0/job/trunk-x86_64/495/console

Making the dependency explicit, by calling ./configure with
--enable-nettle or --disable-nettle depending on the activation status
of package nettle, solves the issue.

For fli4l the fix is still based on buildroot 2018.02.x, but the problem
is the same in recent version of buildroot, and the fix works for master
as well.

Cc: Florian Wolters <florian@florian-wolters.de>
Signed-off-by: Alexander Dahl <post@lespocky.de>