package/fastd: bump version to 22
There's a new feature to offload L2TP to the Linux kernel, which depends
on libmnl.  A new menu option was introduced for that.

It's possible to link an internal version of libmnl
statically for smaller binary size on constraint targets, but this was
not added to the buildroot package for complexity reasons (both build
and licensing).

The COPYING file gained an additional paragraph for internal libmnl
which does not apply, because internal libmnl is not used in the
package.

Link: https://fastd.readthedocs.io/en/latest/releases/v22.html
Signed-off-by: Alexander Dahl <post@lespocky.de>